+++ s4e-web/src/app/components/root/root.component.ts
@@ -1,8 +1,11 @@
+import { filter } from 'rxjs/operators';
 import {Component} from '@angular/core';
 import {resetStores} from '@datorama/akita';
 import {environment} from '../../../environments/environment';
-import {ModalService} from '../../modal/state/modal.service';
 import {NotificationService} from 'notifications';
+import { Router, NavigationEnd } from '@angular/router';
+import { MapService } from 'src/app/views/map-view/state/map/map.service';
+import { SearchResultsService } from 'src/app/views/map-view/state/location-search-results/locations-search-results.service';
 
 @Component({
   selector: 's4e-root',
@@ -12,8 +15,21 @@ import {NotificationService} from 'notifications';
 export class RootComponent {
   PRODUCTION: boolean = environment.production;
 
-  constructor(private modalService: ModalService, private notificationService: NotificationService) {
+  constructor(
+    private notificationService: NotificationService,
+    private _router: Router,
+    private _mapService: MapService,
+    private _locationSearchResultsService: SearchResultsService
+  ) {
     console.log('SOK version ' + environment.version);
+
+    this._router.events
+      .pipe(filter(event => event instanceof NavigationEnd))
+      .subscribe(() => {
+        this._mapService.toggleLoginOptions(false);
+        this._mapService.toggleProductDescription(false);
+        this._locationSearchResultsService.toggleSearchResults(false);
+      });
   }
 
   /**