+++ s4e-web/src/app/views/login/login.module.ts
@@ -1,9 +1,9 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import {LoginComponent} from './login.component';
-import {ReactiveFormsModule} from '@angular/forms';
 import {ShareModule} from '../../common/share.module';
 import {FormErrorModule} from '../../components/form-error/form-error.module';
+import { RecaptchaModule, RecaptchaFormsModule } from 'ng-recaptcha';
 
 @NgModule({
   declarations: [
@@ -12,7 +12,9 @@ import {FormErrorModule} from '../../components/form-error/form-error.module';
   imports: [
     CommonModule,
     FormErrorModule,
-    ShareModule
+    ShareModule,
+    RecaptchaFormsModule,
+    RecaptchaModule,
   ],
   exports: [
     LoginComponent